Menambahkan hak akses dan middleware ke semua pengguna (admin, kasir, owner) dan menambahkan fitur edit profile

+++ database/migrations/2020_04_11_013917_create_detail_transaksis_table.php
@@ -19,7 +19,7 @@ class CreateDetailTransaksisTable extends Migration
             $table->foreign('transaksi_id')->references('id')->on('transaksis')->onDelete('cascade');
             $table->unsignedInteger('produk_id');
             $table->foreign('produk_id')->references('id')->on('produks')->onDelete('cascade');
-            $table->double('quality');
+            $table->bigInteger('quality');
             $table->String('keterangan');
             $table->timestamps();
         });
